pyproject.toml
[tool.poetry]
name = "gehennabot"
version = "0.6.0"
description = ""
authors = ["Claudio Torcato <claudiotorcato@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Pyrogram = "^2.0.106"
python-dotenv = "^1.0.0"
TgCrypto = "^1.2.5"
requests = "^2.31.0"
pyaes = "^1.6.1"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.5.2"
masonite-orm = "^2.22.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
taskipy = "^1.12.0"
blue = "^0.9.1"
isort = "^5.12.0"
ruff = "^0.0.286"


[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.isort]
profile = "black"
line_length = 79

[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff'
format = 'blue .  && isort .'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
